aisaraj-backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
python-dotenv>=1.0
django-filter>=23.5
django-storages>=1.14
boto3>=1.34
google-generativeai>=0.4
gunicorn>=21.2
